google-docs-generator/requirements.txt
# Core dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
rich==13.7.0
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
pathlib==1.0.1

# For document parsing
beautifulsoup4==4.12.2
lxml==4.9.3
pypandoc==1.12

# Note: difflib and hashlib are built-in Python modules

# Utilities
python-dateutil==2.8.2
tabulate==0.9.0
